Refactor feed queries to fix text posts in video feeds
This fixes that regular text posts (i.e. those without embed) appear on
the video-only feeds. It’s done by completely refactoring the feed
queries, so instead of passing HasMedia or HasVideo, we now give a list
of “allowed embeds”, which can be one of:

- `none`: regular text posts
- `image`: posts with images
- `video`: posts with videos

If no allowed embeds are given, it defaults to all (none, image, and
video).